refactor(remediationorchestrator): extract timeout_management.go from reconciler.go
Phase 3 file split: moves timeout default population/validation
(populateTimeoutDefaults, validateControllerTimeouts, getSafeDefaultTimeouts),
effective-timeout resolution (getEffectiveGlobalTimeout, getEffectivePhaseTimeout,
safeFormatTime), timeout detection/handling (checkPhaseTimeouts,
applyInteractiveTimeoutExtension, handlePhaseTimeout), and config validation
(validateTimeoutConfig, getConsecutiveFailureThreshold) into their own file.
Pure structural move: 376 Ginkgo specs still pass.
